% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunctions.R
\name{ParetoForCategoricalVector}
\alias{ParetoForCategoricalVector}
\title{Pareto statistic for categorical data.}
\usage{
ParetoForCategoricalVector(data, normalizeQ = TRUE)
}
\arguments{
\item{data}{A character or factor vector.}

\item{normalizeQ}{Should the Pareto statistic be normalized with the total or not?}
}
\value{
A data frame with columns c( "Item", "ParetoFraction" ).
}
\description{
A synonym/shortcut for \code{\link{ParetoForWeightedItems}}.
}
