% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunctions.R
\name{ParetoForLongForm}
\alias{ParetoForLongForm}
\title{Pareto for long form data.}
\usage{
ParetoForLongForm(data, variables = NULL, normalizeQ = TRUE)
}
\arguments{
\item{data}{A data frame with columns \code{c("Variable", "Item", "Weight")}.}

\item{variables}{The variables to calculate upon.
If NULL then \code{variables = colnames(data)}.}

\item{normalizeQ}{Should the Pareto statistic be normalized with the total or not?}
}
\value{
A data frame with columns \code{c("Variable", "Item", "ParetoFraction")}.
}
\description{
Computes the Pareto statistic for each of the specified columns.
}
\details{
This function works on long form data.
It delegates to the function \code{\link{ParetoForWeightedItems}}.
The data is partitioned according to \code{data$Variable}.
If \code{normalizeQ = TRUE} then for each part \code{data2} the Pareto statistic is computed with the formula:
\code{ cumsum(sort(data2$Weight, decreasing=TRUE)) / sum(data2$Weight) }.
}
