% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunctions.R
\name{ParetoForNumericalVector}
\alias{ParetoForNumericalVector}
\title{Pareto statistic for numerical vector}
\usage{
ParetoForNumericalVector(data, normalizeQ = TRUE)
}
\arguments{
\item{data}{A numeric vector.}

\item{normalizeQ}{Should the Pareto statistic be normalized with the total or not?}
}
\value{
A numeric vector.
}
\description{
Pareto statistic for numerical vector
}
\details{
The simple formula
\code{cumsum(rev(sort(data))) / sum(data)}
is used.
}
