% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunctions.R
\name{ParetoForWeightedItems}
\alias{ParetoForWeightedItems}
\title{Pareto statistic for weighted items.}
\usage{
ParetoForWeightedItems(data, normalizeQ = TRUE)
}
\arguments{
\item{data}{A data frame with columns "Item" and "Weight", a numerical vector,
or a categorical vector.}

\item{normalizeQ}{Should the Pareto statistic be normalized with the total or not?}
}
\value{
A data frame with columns \code{c("Item", "ParetoFraction")}.
}
\description{
Compute the Pareto position and fraction for each weighted item.
The items are sorted in descending order according to their weight.
(That gives the Pareto positions.)
The Pareto fractions are derived as ratios between the cumulative sum of the weights
and the total weight.
}
\details{
If \code{data} is a character vector then the weights of the elements are assumed 1.
If \code{data} is a numerical vector with named elements that vector is turned into
a item-weight data frame.
If \code{data} is a numerical vector without named elements that vector is turned into
a item-weight data frame with the items being vector's indexes.
}
