% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunctions.R
\name{ParetoFramePlot}
\alias{ParetoFramePlot}
\title{Plot points with a Pareto frame.}
\usage{
ParetoFramePlot(data, main = NULL, xlab = "Index",
  ylab = "ParetoFraction", xFraction = 0.2, yFraction = 0.8,
  showParetoTicksQ = TRUE, ...)
}
\arguments{
\item{data}{A numerical vector.}

\item{main}{A string for the title of the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{xFraction}{The Pareto fraction for the x-axis.}

\item{yFraction}{The Pareto fraction for the y-axis.}

\item{showParetoTicksQ}{Should the Pareto ticks be shown or not?}

\item{...}{Additional parameters for \code{\link{base::plot}}.}
}
\description{
Plot the points of a vector and overlay a Pareto grid
(without computing the Pareto statistic.)
}
\details{
This plot can be used when the Pareto statistic is
computed with some other functions.
(Or for for ad hoc Pareto computations.)
}
\seealso{
Other Pareto plots: \code{\link{ParetoPlot}}
}
\concept{Pareto plots}
