% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoLawFunctions.R
\name{ParetoItems}
\alias{ParetoItems}
\title{Top Pareto items determination.}
\usage{
ParetoItems(data, colName, paretoFraction)
}
\arguments{
\item{data}{A data frame.}

\item{colName}{The column name of items to be counted.}

\item{paretoFraction}{A number between 0 and 1 specifying the Pareto fraction.}
}
\value{
A data frame with columns c( "Item", "Score", "ParetoFraction" ).
}
\description{
Finds the Pareto items in a column of a data frame.
}
\seealso{
Other Pareto items: \code{\link{ParetoPositions}}
}
\concept{Pareto items}
