% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoLawFunctions.R
\name{ParetoLawBasePlot}
\alias{ParetoLawBasePlot}
\title{Pareto for a numerical vector.}
\usage{
ParetoLawBasePlot(dataVec, main = NULL, xlab = "|levels|",
  ylab = "\%", xFraction = 0.2, yFraction = 0.8, ...)
}
\arguments{
\item{dataVec}{A numerical vector.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{plotTitle}{A string for the title of the plot.}
}
\description{
Make a plot that demonstrates the adherence to Pareto Principle of
a numerical vector with contingency values (counts).
}
\details{
Pareto plots
}
