% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoLawFunctions.R
\name{ParetoLawData}
\alias{ParetoLawData}
\title{Pareto statistics computation.}
\usage{
ParetoLawData(dataVec, weightsVec = NULL)
}
\arguments{
\item{dataVec}{A vector with categorical entries.}

\item{weightVec}{A vector of weights for the values of \code{dataVec};
see the argument wt_var of \code{plyr::count}.}
}
\description{
Sorts the tally of given categorical data descendingly
and computes the list of the cumulative sums.
}
