% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoLawFunctions.R
\name{ParetoLawDataForColumns}
\alias{ParetoLawDataForColumns}
\title{Pareto statistics for specified columns.}
\usage{
ParetoLawDataForColumns(data, columnNames)
}
\arguments{
\item{data}{A data frame.}

\item{columnNames}{A list of column names corresponding to categrical columns in \code{data}.}
}
\description{
Apply \code{\link{ParetoLawData}} over a list of names.
}
