% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoLawFunctions.R
\name{ParetoLawForCountsBasePlot}
\alias{ParetoLawForCountsBasePlot}
\title{Pareto for a categorical vector.}
\usage{
ParetoLawForCountsBasePlot(dataVec, main = NULL, xlab = "|levels|",
  ylab = "\%", ...)
}
\arguments{
\item{dataVec}{A vector with categorical entries.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{plotTitle}{A string for the title of the plot.}
}
\description{
Make a plot that demonstrates the adherence to Pareto Principle
of the unique entries of a categorical vector.
}
\concept{Pareto plots}
