% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoLawFunctions.R
\name{ParetoLawPlot}
\alias{ParetoLawPlot}
\title{Pareto plot for name-value pairs.}
\usage{
ParetoLawPlot(data, normalizeQ = TRUE, separatedPlotsQ = TRUE,
  main = NULL, scales = "fixed", nrow = NULL, ncol = NULL, ...)
}
\arguments{
\item{data}{A two column data frame with the first column of a variable name
and second column a numerical vector \code{c("Name", "Value")}.}

\item{normalizeQ}{Should the Pareto cumulative values be normalized or not?}

\item{separatedPlotsQ}{Should the plotted Pareto curves be separated or not?}

\item{main}{A string for the title of the plot.}

\item{scales}{Sames as \code{scales} of \code{ggplot2::facet_wrap}.}

\item{nrow}{Sames as \code{nrow} of \code{ggplot2::facet_wrap}.}

\item{ncol}{Sames as \code{ncol} of \code{ggplot2::facet_wrap}.}
}
\description{
Make a plot that demonstrates the adherence to the Pareto law of a
a data frame with item names and weight values.
}
\details{
The data frame \code{data} is split by \code{data$Name}.
The column \code{data$Value} is used to make the Pareto Principle curves.
Each plot has percentage vertical lines.

Pareto plots
}
