% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoLawFunctions.R
\name{ParetoLawPlotForColumns}
\alias{ParetoLawPlotForColumns}
\title{Pareto Law plots over a data frame columns.}
\usage{
ParetoLawPlotForColumns(data, columnNames = colnames(data),
  weights = NULL, ...)
}
\arguments{
\item{data}{A data frame.}

\item{columnNames}{A list of column names corresponding to categorical columns in \code{data}.}

\item{...}{Arguments for \code{ggplot2::facet_wrap}.}
}
\description{
Apply \code{\link{ParetoLawData}} function over a list of names
and make a multi-panel ggplot.
}
\details{
This function make the Pareto plots over a wide form data frame.
The function \code{\link{ParetoLawPlot}} makes the plots over long form data.

Pareto plots
}
