% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunctions.R
\name{ParetoPlot}
\alias{ParetoPlot}
\title{Pareto plot for a vector.}
\usage{
ParetoPlot(data, ...)
}
\arguments{
\item{data}{A character, factor, or numerical vector.}

\item{...}{Additional parameters for \code{\link{ParetoPlotForVariables}}.}
}
\description{
Computes the Pareto statistic for a vector and plots it.
}
\details{
This function is a shortcut to \code{\link{ParetoPlotForVariables}}.
}
\seealso{
Other Pareto plots: 
\code{\link{ParetoFramePlot}()}
}
\concept{Pareto plots}
