% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunctions.R
\name{ParetoPlotForColumns}
\alias{ParetoPlotForColumns}
\title{Pareto plots over the columns of a data frame.}
\usage{
ParetoPlotForColumns(data, columnNames = colnames(data),
  normalizeQ = TRUE, separatedPlotsQ = TRUE, main = NULL,
  scales = "fixed", nrow = NULL, ncol = NULL, ...)
}
\arguments{
\item{data}{A data frame.}

\item{columnNames}{A list of column names corresponding to categorical columns in \code{data}.}

\item{normalizeQ}{Should the Pareto cumulative values be normalized or not?}

\item{separatedPlotsQ}{Should the plotted Pareto curves be separated or not?}

\item{main}{A string for the title of the plot.}

\item{scales}{Sames as \code{scales} of \code{ggplot2::facet_wrap}.}

\item{nrow}{Sames as \code{nrow} of \code{\link{ggplot2::facet_wrap}}.}

\item{ncol}{Sames as \code{ncol} of \code{\link{ggplot2::facet_wrap}}.}

\item{...}{Parameters for \code{\link{ggplot2::theme}}}
}
\description{
Apply \code{\link{ParetoLawData}} function over a list of names
and make a multi-panel ggplot.
}
\details{
This function make the Pareto plots over a wide form data frame.
The function \code{\link{ParetoPlotForVariables}} makes the plots over long form data.

Pareto plots
}
