% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunctions.R
\name{ParetoPlotForVariables}
\alias{ParetoPlotForVariables}
\title{Pareto plot for variables.}
\usage{
ParetoPlotForVariables(data, normalizeQ = TRUE, separatedPlotsQ = TRUE,
  main = NULL, scales = "fixed", nrow = NULL, ncol = NULL, ...)
}
\arguments{
\item{data}{A two column data frame with the first column of a variable name
and second column a numerical vector \code{c("Variable", "Value")}.}

\item{normalizeQ}{Should the Pareto cumulative values be normalized or not?}

\item{separatedPlotsQ}{Should the plotted Pareto curves be separated or not?}

\item{main}{A string for the title of the plot.}

\item{scales}{Sames as \code{scales} of \code{ggplot2::facet_wrap}.}

\item{nrow}{Sames as \code{nrow} of \code{\link{ggplot2::facet_wrap}}.}

\item{ncol}{Sames as \code{ncol} of \code{\link{ggplot2::facet_wrap}}.}

\item{...}{Parameters for \code{\link{ggplot2::theme}}}
}
\description{
Make a plot that demonstrates the adherence to the Pareto Principle
of a data frame with item names and weight values.
}
\details{
The data frame \code{data} is split by \code{data$Variable}.
The column \code{data$Value} is used to make the Pareto Principle curves.
Each plot has percentage vertical lines.

Pareto plots
}
