% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoLawFunctions.R
\name{ParetoPositions}
\alias{ParetoPositions}
\title{Positions of top Pareto items determination.}
\usage{
ParetoPositions(dataVec, paretoFraction = 1)
}
\arguments{
\item{dataVec}{A data vector.}

\item{paretoFraction}{A number between 0 and 1 specifying the Pareto fraction.}
}
\value{
A data frame with columns c( "Index", "Score", "ParetoFraction" ).
}
\description{
Finds the positions of the Pareto items in a column of a data frame.
}
\seealso{
Other Pareto items: \code{\link{ParetoItems}}
}
\concept{Pareto items}
