% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunctions.R
\name{ParetoForVariables}
\alias{ParetoForVariables}
\title{Pareto for variables.}
\usage{
ParetoForVariables(data, columnNames = colnames(data), normalizeQ = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{columnNames}{The column names of the variables.}

\item{normalizeQ}{Should the Pareto statistic be normalized with the total or not?}
}
\value{
A data frame with columns \code{c("Variable", "Item", "ParetoFraction")}.
}
\description{
Computes the Pareto statistic for each of the specified columns.
}
\details{
This function works on wide form data.
(The function \code{\link{ParetoForWeightedItems}} works on long form data.)
Note that the specified columns can be a mix of categorical and numerical columns.
}
