% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoFunctions.R
\name{ParetoPlot}
\alias{ParetoPlot}
\title{Pareto plot for a vector.}
\usage{
ParetoPlot(data)
}
\arguments{
\item{data}{A character, factor, or numerical vector.}
}
\description{
Pareto plot for a vector.
}
